+++ compatibility/chat-commands-lib.js
@@ -1,5 +1,24 @@
 Hooks.on('chatCommandsReady', function (chatCommands) {
-	// (GM Only) This Command will display the text after the command as well as invoke the method
+	chatCommands.registerCommand(
+		chatCommands.createCommandFromData({
+			commandKey: '/as',
+			invokeOnCommand: (chatlog, messageText, chatdata) => NarratorTools.chatMessage.narrate(messageText),
+			shouldDisplayToChat: false,
+			iconClass: 'fa-sticky-note',
+			description: 'Makes the next messages be sent as the [speaker]. "/as" resets.',
+			gmOnly: true,
+		})
+	);
+	chatCommands.registerCommand(
+		chatCommands.createCommandFromData({
+			commandKey: '/describe',
+			invokeOnCommand: (chatlog, messageText, chatdata) => NarratorTools.chatMessage.narrate(messageText),
+			shouldDisplayToChat: false,
+			iconClass: 'fa-sticky-note',
+			description: 'Display a description in chat',
+			gmOnly: true,
+		})
+	);
 	chatCommands.registerCommand(
 		chatCommands.createCommandFromData({
 			commandKey: '/narrate',
@@ -12,11 +31,11 @@ Hooks.on('chatCommandsReady', function (chatCommands) {
 	);
 	chatCommands.registerCommand(
 		chatCommands.createCommandFromData({
-			commandKey: '/describe',
+			commandKey: '/note',
 			invokeOnCommand: (chatlog, messageText, chatdata) => NarratorTools.chatMessage.narrate(messageText),
 			shouldDisplayToChat: false,
 			iconClass: 'fa-sticky-note',
-			description: 'Display a description in chat',
+			description: 'Display a note only visible to the DM',
 			gmOnly: true,
 		})
 	);